working split version
