+++ lib/_enums/score_is.dart
@@ -1,0 +1,1 @@
+enum ScoreIs { INCREMENTING, INCREMENTED, UNCHANGING }
