+++ lib/_enums/timer_state.dart
