+++ lib/main_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../enums/timer_state.dart';
+import '_enums/timer_state.dart';
 import 'dart:ui';
 import 'dart:async';
 import 'atoms/score.dart';
@@ -8,7 +8,7 @@ class _MainPageState extends State<MainPage> with TickerProviderStateMixin {
   int _counter = 0;
   int _left_count = 0;
   int _right_count = 0;
-  TimerState _timer_state = TimerState.paused;
+  TimerIs _timer_state = TimerIs.paused;
   Duration _max_time = Duration(minutes: 0, seconds: 20);
   late Timer _timer;
   Stopwatch _stopwatch = Stopwatch();
@@ -21,19 +21,19 @@ class _MainPageState extends State<MainPage> with TickerProviderStateMixin {
         vsync: this, duration: const Duration(milliseconds: 250));
   }
 
-  IconData getTimerIconForTimerState(TimerState timer_state) {
+  IconData getTimerIconForTimerState(TimerIs timer_state) {
     IconData ret_icon;
     switch (timer_state) {
-      case TimerState.none:
+      case TimerIs.none:
         ret_icon = Icons.question_mark;
         break;
-      case TimerState.finished:
+      case TimerIs.finished:
         ret_icon = Icons.restore;
         break;
-      case TimerState.paused:
+      case TimerIs.paused:
         ret_icon = Icons.play_arrow;
         break;
-      case TimerState.running:
+      case TimerIs.running:
         ret_icon = Icons.pause;
         break;
       default:
@@ -94,11 +94,11 @@ class _MainPageState extends State<MainPage> with TickerProviderStateMixin {
   /* Pauses timer if and only if it is running */
   void safePauseTimer() {
     setState(() {
-      if (_timer_state != TimerState.running) {
+      if (_timer_state != TimerIs.running) {
         return;
       }
       myTimerAnimation.reverse();
-      _timer_state = TimerState.paused;
+      _timer_state = TimerIs.paused;
       _stopwatch.stop();
       _timer.cancel();
     });
@@ -107,16 +107,16 @@ class _MainPageState extends State<MainPage> with TickerProviderStateMixin {
   void _handleTimer() {
     setState(() {
       switch (_timer_state) {
-        case TimerState.none:
+        case TimerIs.none:
           break;
-        case TimerState.finished:
-          _timer_state = TimerState.paused;
+        case TimerIs.finished:
+          _timer_state = TimerIs.paused;
           _stopwatch.reset();
           _timer.cancel();
           break;
-        case TimerState.paused:
+        case TimerIs.paused:
           myTimerAnimation.forward();
-          _timer_state = TimerState.running;
+          _timer_state = TimerIs.running;
           _stopwatch.start();
           Duration period = Duration(seconds: 1);
           if (getElapsedFromMaxTime() <= Duration(seconds: 10)) {
@@ -131,7 +131,7 @@ class _MainPageState extends State<MainPage> with TickerProviderStateMixin {
                   timer.cancel();
                   setState(() {
                     _stopwatch.stop();
-                    _timer_state = TimerState.finished;
+                    _timer_state = TimerIs.finished;
                   });
                 }
                 setState(() {});
@@ -141,7 +141,7 @@ class _MainPageState extends State<MainPage> with TickerProviderStateMixin {
             setState(() {});
           });
           break;
-        case TimerState.running:
+        case TimerIs.running:
           safePauseTimer();
           break;
         default:
